Adapt corert to ConditionalWeakTable moving to shared
